Move all NUS APIs under api/nus
